refactor: update structure to DDD architecture
