+++ src/Domain.Endpoint/dtos/chicken.dto.ts
