+++ src/Domain.Endpoint/dtos/lote.dto.ts
