+++ src/Domain.Endpoint/dtos/salud.dto.ts
