+++ src/Domain.Endpoint/models/alert.model.ts
