+++ src/Domain.Endpoint/models/base.model.ts
