+++ src/Domain.Endpoint/models/chicken.model.ts
