+++ src/Domain.Endpoint/models/health.enum.ts
