+++ src/Domain.Endpoint/models/lote.model.ts
