+++ src/Domain.Endpoint/models/report.model.ts
