+++ src/Domain.Endpoint/models/salud.model.ts
