+++ src/Infrastructure.Endpoint/builders/sqlOperations.enum.ts
