+++ src/WebApi/routes/lotes.routes.ts
