+++ src/app.ts
@@ -1,11 +1,11 @@
 import express from 'express';
 import swaggerUI from "swagger-ui-express";
-import specs from "./swagger/swagger";
-import chickenRoutes from "./routes/chickens.routes"
-import lotesRoutes from "./routes/lotes.routes";
-import saludRoutes from "./routes/salud.routes";
-import alertsRoutes from "./routes/alerts.routes";
-import reportesRoutes from "./routes/reportes.routes";
+import specs from "./WebApi/swagger/swagger";
+import chickenRoutes from "./WebApi/routes/chickens.routes"
+import lotesRoutes from "./WebApi/routes/lotes.routes";
+import saludRoutes from "./WebApi/routes/salud.routes";
+import alertsRoutes from "./WebApi/routes/alerts.routes";
+import reportesRoutes from "./WebApi/routes/reportes.routes";
 
 const app = express();
 const PORT=3000;
